requirements.txt
# Core cryptography
liboqs-python>=0.14.0

# Web framework for OIDC server
Flask==3.0.0
Flask-CORS==4.0.0

# JWT and OAuth support
PyJWT==2.8.0
authlib==1.3.0

# Cryptography utilities
cryptography==42.0.0

# HTTP client
requests==2.31.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Benchmarking
numpy==1.26.0
matplotlib==3.8.0
